requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

langchain==0.2.12
langchain-openai==0.1.20
langchain-community==0.2.11

faiss-cpu==1.8.0.post1
numpy==1.26.4
tiktoken==0.7.0
